holy shit completely scalable map
reworked the splitting/seeding so that quads are pre-split to avoid huge
runtime inserts as the quad size scales up when moving the map fuirther
out. Now 1) quads are split preemptively (expecting to need to do
so..duh) and 2) only one quad is seeded on each tick. This means that on
a given tick only true leaves (ones that can't be split further) have
deferred inserts, making the number of spawned sprites linear.